Improve error type extraction for logs
1. Port the type unwrapping code from our metrics to our util package
   and use it
2. Rename tag.ErrorType to tag.ServiceErrorType and create a new
   tag.ErrorType
3. Report the type for all task errors we log
