+++ client/frontend/metric_client.go
@@ -84,7 +84,7 @@ func (c *metricClient) finishMetricsRecording(
 			*serviceerror.WorkflowExecutionAlreadyStarted:
 			// noop - not interest and too many logs
 		default:
-			c.throttledLogger.Info("frontend client encountered error", tag.Error(err), tag.ErrorType(err))
+			c.throttledLogger.Info("frontend client encountered error", tag.Error(err), tag.ServiceErrorType(err))
 		}
 		metrics.ClientFailures.With(handler).Record(1, metrics.ServiceErrorTypeTag(err))
 	}
